Added help texts (#68)
* Incorrect Type help text

First attempt at authoring a help message. Is this level of detail
correct? Should an example be provided?

* Help text for Incorrect Dimensions

Incorrect Dimensions :  The row dimensions do not match the header
dimensions

* Required field help text

Assume changing the data to a “missing value” is not a valid solution
for this error.

* add quotes to "required"

* Empty rows help text

* Incorrect headers help text

* unique constraint help text

* Things you can try:

* Things you can try:

* grammar

* check schema message

* Things you can try:

* Things you can try:

* Things you can try:

* Things you can try:

* This row is empty.

revert schema_005.md changed incorrectly

* revert to placeholder message

changed this instead of structure_005.md

* formatting

* This row has a different number of columns compared to the header

Same message as schema_002.md

* Remaining Structure errors